voice end task fix

+++ app/BUILD/src/DeskPRO/Bundle/AppBundle/Twilio/TwilioAdapter.php
@@ -911,7 +911,9 @@ class TwilioAdapter
             throw $e;
         }
 
-        $task->delete();
+        if ($task->sid) {
+            $task->delete();
+        }
     }
 
     /**
